feat: Add YouTube playlist import and song lyrics offset
Introduces the ability to import entire YouTube playlists by adding a new tab to the import modal and a `fetchPlaylistItems` function.

Also, enhances song playback by adding a `lyricsOffset` property to the `Song` type. This allows for fine-grained synchronization adjustments of lyrics displayed during playback, improving the user experience for songs with misaligned lyrics.

Includes UI improvements for the Now Playing bar, such as adding a shuffle button and refining the progress bar design. Adds drag-and-drop functionality to the song list items for playlist reordering.